fix(conversation):#WB-3835 rename  component MessagePreview as best practices
