+++ conversation/frontend/src/features/message-list/components/MessagePreview/components/RecipientAvatar.tsx
@@ -1,8 +1,8 @@
 import { Avatar } from '@edifice.io/react';
 import { IconGroupAvatar, IconQuestionMark } from '@edifice.io/react/icons';
 import { useTranslation } from 'react-i18next';
-import { useRecipientAvatar } from '~/features/message-list/components/message-preview/hooks/useRecipientAvatar';
 import { Recipients } from '~/models';
+import { useRecipientAvatar } from '../hooks/useRecipientAvatar';
 
 interface RecipientAvatarProps {
   recipients: Recipients;
