+++ conversation/frontend/src/features/message-list/components/MessagePreview/components/SenderAvatar.tsx
