+++ conversation/frontend/src/features/message-list/components/MessagePreview/components/UserFolderIcon.tsx
