+++ conversation/frontend/src/features/message-list/components/MessagePreview/hooks/useRecipientAvatar.ts
